requirements.txt
requests==2.31.0
feedparser==6.0.12
beautifulsoup4==4.12.2
lxml>=4.9.0
python-dotenv==1.0.0